Hazelcast 3.6: New approach for capturing hazelcast instance name
Instrument ClientInvocation to capture Hazelcast client instance instead of having to use context store
