+++ dd-java-agent/instrumentation/hazelcast/hazelcast-3.6/src/main/java/datadog/trace/instrumentation/v3/DistributedObjectDecorator.java
@@ -95,6 +95,10 @@ public class DistributedObjectDecorator extends ClientDecorator {
 
   public AgentSpan onHazelcastInstance(final AgentSpan span, HazelcastInstance instance) {
 
+    if (span == null) {
+      return span;
+    }
+
     if (instance != null
         && instance.getLifecycleService() != null
         && instance.getLifecycleService().isRunning()) {
